Game에 구현된 split 기능을 모듈화
